backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
redis==6.4.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
alembic==1.16.5
requests==2.32.5

# Data extraction dependencies
pystac-client==0.8.1
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
fiona==1.9.5
pandas==2.1.4
numpy==1.24.4
pydantic==2.5.2
httpx==0.25.2
aiofiles==23.2.1

# Additional dependencies for impact simulation
netCDF4==1.6.5
pyproj==3.6.1
cartopy==0.22.0
matplotlib==3.8.2
scipy==1.11.4
xarray==2023.12.0